+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/action/ReconnectAction.kt
@@ -1,12 +1,22 @@
 package io.github.pandier.intellijdiscordrp.action
 
+import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.application.EDT
 import com.intellij.openapi.project.DumbAwareAction
-import io.github.pandier.intellijdiscordrp.service.discordService
+import io.github.pandier.intellijdiscordrp.service.DiscordService
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 class ReconnectAction : DumbAwareAction() {
 
+    override fun getActionUpdateThread(): ActionUpdateThread =
+        ActionUpdateThread.EDT
+
     override fun actionPerformed(e: AnActionEvent) {
-        discordService.reconnect()
+        val discordService = DiscordService.getInstance()
+        discordService.scope.launch(Dispatchers.EDT) {
+            discordService.reconnect()
+        }
     }
 }