+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/listener/RichPresenceProjectListener.kt
@@ -1,27 +1,35 @@
 package io.github.pandier.intellijdiscordrp.listener
 
+import com.intellij.openapi.application.EDT
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.project.ProjectManagerListener
 import com.intellij.openapi.startup.StartupActivity
 import io.github.pandier.intellijdiscordrp.DiscordRichPresencePlugin
 import io.github.pandier.intellijdiscordrp.activity.ActivityContext
-import io.github.pandier.intellijdiscordrp.service.discordService
-import io.github.pandier.intellijdiscordrp.service.timeTrackingService
+import io.github.pandier.intellijdiscordrp.service.DiscordService
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 
 class RichPresenceProjectListener : ProjectManagerListener, StartupActivity {
 
     override fun projectClosed(project: Project) {
-        timeTrackingService.stop(project)
+        val discordService = DiscordService.getInstance()
         if (discordService.activityContext?.project?.get() == project) {
-            discordService.clearActivity()
+            discordService.scope.launch(Dispatchers.EDT) {
+                discordService.clearActivity()
+            }
         }
     }
 
     override fun runActivity(project: Project) {
         DiscordRichPresencePlugin.registerEditorListeners()
 
+        val discordService = DiscordService.getInstance()
         if (discordService.activityContext?.project?.get() != project) {
-            discordService.changeActivity(ActivityContext.create(project = project))
+            val context = ActivityContext.create(project = project)
+            discordService.scope.launch(Dispatchers.EDT) {
+                discordService.changeActivity(context)
+            }
         }
     }
 }