+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/service/TimeTrackingService.kt
@@ -5,28 +5,12 @@ import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import java.time.Instant
 
-val timeTrackingService: TimeTrackingService
-    get() = service()
-
-@Service
+@Service(Service.Level.PROJECT)
 class TimeTrackingService {
-    private val times: MutableMap<String, Instant> = mutableMapOf()
-
-    fun track(
-        project: Project,
-        time: Instant = Instant.now()
-    ): Instant {
-        times[project.name] = time
-        return time
+    companion object {
+        @JvmStatic
+        fun getInstance(project: Project): TimeTrackingService = project.service()
     }
 
-    fun stop(project: Project) {
-        times.remove(project.name)
-    }
-
-    fun get(project: Project): Instant? =
-        times[project.name]
-
-    fun getOrTrack(project: Project): Instant =
-        get(project) ?: track(project, Instant.now())
+    val start: Instant = Instant.now()
 }